Adding GeoDB token to trust wallet
This is the addition of GeoDB token to Trust wallet. Our app Geo cash had over  120000 downloads in the last 4 months and a lot of our users use Trust wallet!
We were requested to add Geo to the trust wallet by our users!

This folder contains the GeoDB coin logo in png format.